orchestrator_agent/pyproject.toml
[project]
name = "orchestrator-agent-a2a"
version = "0.1.0"
description = "Orchestrator agent with A2A server support"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.8.0",
    "a2a-sdk>=0.3.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
